Show no account item when spinner is empty
Show "reddit by brian" in the title and one item that says "No account"
in the spinner when there are no accounts. This is cleaner than the
prior way of making a custom view that swapped between two views.

Had to add a title back to the activity or it doesn't appear in the
recents panel...