Fix #35: show points when no voting arrows
Stop using a format string, since it is too difficult to juggle around
all the optional items. This is no doubt bad for translation, but I
don't expect there will ever be translations...

Take the opportunity to reuse a StringBuilder to avoid extra allocations
when the View is reused in a list.

Also apply the same fixes to the long details text and fix the plural
strings for ups and downs.