Fix #135: handle thing links even better
This fixes the thing pager being empty in landscape mode. A previous
hack to reset the ThingPager in a separate runnable was erasing the
pager. The approach in this change expands an existing method to
setup the thing pager in one transaction. This also means pressing back
will go back to the calling application which is nice. The remaining
problem with the up button doing nothing in this case will be revisited
in issue #134.