Issue #204: add rudimentary code block formatting
Recognize indented lines and make them code blocks. This creates a new
TypefaceSpan for each line, so it could be improved to apply only a
single span for multiple lines to avoid allocating unnecessary memory.

Furthermore, the Formatter will still apply formatting to the content
that was placed in a code block.

These two issues need to be fixed before marking this issue as fixed.