Issue #205: remove shrinking and other fixes
Remove the shrinking animation, because our db operations are
asynchronous and could finish well after the view is snapped back to its
original height. Since we're asynchronous just reset the view to its
original state when binding new data.

Remove the newVelocityTracker method in favor of just obtaining and
recycling. It seems like clearing the data doesn't work for some reason,
so this seems to fix the oversensitive dismissing. Change the condition
to match Roman's code sample.