Issue #205: eliminate thumb flicker and more
Eliminate the thumbnail flicker by only animating the thumbnail alpha
when a thumbnail is set by the AsyncTask. If we already have a thumbnail
sitting in the cache, then don't animate it to avoid flicker when items
are shifted in response to dismissing something.

Hook up the listener in ThingListFragment to hide things.